% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bilby}
\alias{bilby}
\title{Bilby example data}
\format{
A data frame with 3000 rows and 10 variables.
}
\source{
Simulated data.
}
\usage{
bilby
}
\description{
A simulated 10-dimensional dataset containing a bilby-shaped pattern
in the \code{x3} and \code{x8} variables. The remaining variables are noise.
}
\examples{
data("bilby", package = "bilbyr")
bilby |>
slice_head(n = 1500) |>
 ggplot(aes(x3, x8)) +
 geom_point(size = 0.5) +
 theme_minimal()

}
\keyword{datasets}
