% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_stringy}
\alias{simple_stringy}
\title{Simple stringy example data}
\format{
A data frame with 500 rows and 4 variables.
}
\source{
Simulated data.
}
\usage{
simple_stringy
}
\description{
A simulated 4-dimensional dataset with a sine-shaped pattern
in \code{x1} and \code{x2}. Variables \code{x3} and \code{x4} are independent
Gaussian noise.
}
\keyword{datasets}
